+++ Classes/Hooks/PageRenderer/TrackingCodeInjector.php
@@ -59,7 +59,7 @@ final class TrackingCodeInjector
         $this->scriptTagBuilder->setRequest($request);
 
         $scriptCode = $this->javaScriptTrackingCodeBuilder->getTrackingCode();
-        if ($configuration->tagManagerContainerId !== []) {
+        if ($configuration->tagManagerContainerIds !== []) {
             $scriptCode .= $this->tagManagerCodeBuilder->getCode();
         }